Fix rx->nr integer overflow for large VPC output and latent accumulator overflow
rx->nr (output row count) is computed as nobs * nsim. In VPC runs with many
subjects, many observation timepoints, and many simulations (e.g., 1000
subjects x 2148 obs x 1000 VPC sims -> 2.148B rows > INT_MAX), this
previously overflowed to a negative int, causing a crash at vector allocation.

Changes:
- rx_solve::nr: int -> int64_t
- rxode2_df.cpp: compute nr using int64_t arithmetic; add overflow guard that
  throws an informative error ("too large") before any allocation is attempted;
  add (R_xlen_t) casts at all allocation call sites; fix negation in compact
  row.names representation
- rxData.cpp: add (R_xlen_t) cast to NumericMatrix constructor using rx->nr
- rxData.cpp: curSolve, curEvent, eLen -> int64_t to prevent latent overflow
  in the subject-setup loop when neq * total_n_all_times > INT_MAX

Test: rxSolve nr overflow guard gives informative error not segfault

+++ inst/include/rxode2parseStruct.h
@@ -250,7 +250,7 @@ typedef struct {
   int nevid9;
   int nobs; // isObs() observations
   int nobs2; // evid=0 observations
-  int nr;
+  int64_t nr; // int64_t because nobs*nsim can exceed INT_MAX in VPC with many subjects/timepoints/sims
   int add_cov;
   int matrix;
   int needSort;
